updated file directory temp

+++ routes/index.js
@@ -366,11 +366,13 @@ function deleteAllTempFiles(directory){
       if (err) throw err;
 
       for (const file of files) {
-        fs.unlink(path.join(directory, file), err => {
-          if (err) throw err;
-
-          console.log("done!!");
-        });
+        if(file !== "test.jpg"){
+          fs.unlink(path.join(directory, file), err => {
+            if (err) throw err;
+  
+            console.log("done!!");
+          });
+        }
       }
     });
 }
